getAsXhtml did not merge parser and visitor errors.
Whenever the errorReporting option only included `E_PARSE & E_ERROR`
it was impossible to get the errors generated by the visitor and parser 
as these were not propagated. This commit changes that by merging the
parser and visitor errors into this error listing.

This is a replica of the behaviour in the getAsDocbook() method.